Fixing specs & fixtures (moved)
