+++ fixtures/vcr_cassettes/ChartMogul_Customer/API_Interactions/adds_required_tags.yml
