+++ fixtures/vcr_cassettes/ChartMogul_Customer/API_Interactions/merges_customers.yml
