+++ fixtures/vcr_cassettes/ChartMogul_Customer/API_Interactions/removes_tags.yml
