+++ fixtures/vcr_cassettes/ChartMogul_Customer/API_Interactions/updates_customer.yml
